+++ ui/src/app/dashboard/components/panel/dashboard-panel.component.html
@@ -30,6 +30,7 @@
                     fxFlex="100"
                     [dashboard]="dashboard"
                     [editMode]="editMode"
+                    [createMode]="createMode"
                     [(viewMode)]="viewMode"
                     [hasDashboardWritePrivileges]="hasDashboardWritePrivileges"
                     [timeRangeVisible]="timeRangeVisible"
