+++ ui/src/app/dashboard/components/panel/dashboard-toolbar/dashboard-toolbar.component.ts
@@ -72,6 +72,9 @@ export class DashboardToolbarComponent {
     hasDashboardWritePrivileges: boolean;
 
     @Input()
+    createMode: boolean;
+
+    @Input()
     timeSettings: TimeSettings;
 
     @Output()
